moving size_seq_satn tests
